code adjustment for mov. from arduino mega to due

+++ include/UI/lib/BitMapButton.hpp
@@ -10,7 +10,7 @@ namespace UI
     public:
         BitMapButton(WindowBase *_parent, position _upper_left,
                      uint16_t _bitmap_width, uint16_t _bitmap_height, CALL_OBJECT_TYPE *_call_object,
-                     const unsigned int *_on_touch_bitmap, const unsigned int *_on_release_bitmap,
+                     const unsigned short *_on_touch_bitmap, const unsigned short *_on_release_bitmap,
                      uint8_t _border_weight = 3,
                      uint8_t _border_to_bitmap_gap = 5) : ButtonBase<CALL_OBJECT_TYPE>(_parent, _upper_left, {_upper_left.x_pos + _bitmap_width - 1 + _border_weight * 2 + _border_to_bitmap_gap * 2, _upper_left.y_pos + _bitmap_height - 1 + _border_weight * 2 + _border_to_bitmap_gap * 2}, _call_object, _border_weight), // -1 we start counting the pixels at 0...
                                                           on_touch_bitmap(_on_touch_bitmap), on_release_bitmap(_on_release_bitmap), border_to_bitmap_gap(_border_to_bitmap_gap)
@@ -28,8 +28,8 @@ namespace UI
         void _draw_touched_content() override;
 
     private:
-        const unsigned int *on_touch_bitmap;
-        const unsigned int *on_release_bitmap;
+        const unsigned short *on_touch_bitmap;
+        const unsigned short *on_release_bitmap;
 
         uint8_t border_to_bitmap_gap; // in pixels
     };
@@ -47,7 +47,7 @@ void UI::BitMapButton<CALL_OBJECT_TYPE>::_draw_released_content()
     this->display->drawBitmap(this->upper_left.x_pos + this->get_border_weight() + this->border_to_bitmap_gap,
                               this->upper_left.y_pos + this->get_border_weight() + this->border_to_bitmap_gap,
                               bitmap_width, bitmap_height,
-                              const_cast<unsigned int *>(this->on_touch_bitmap), 1);
+                              const_cast<unsigned short *>(this->on_touch_bitmap), 1);
 }
 
 template <typename CALL_OBJECT_TYPE>
@@ -62,5 +62,5 @@ void UI::BitMapButton<CALL_OBJECT_TYPE>::_draw_touched_content()
     this->display->drawBitmap(this->upper_left.x_pos + this->get_border_weight() + this->border_to_bitmap_gap,
                               this->upper_left.y_pos + this->get_border_weight() + this->border_to_bitmap_gap,
                               bitmap_width, bitmap_height,
-                              const_cast<unsigned int *>(this->on_touch_bitmap), 1);
+                              const_cast<unsigned short *>(this->on_touch_bitmap), 1);
 }
